backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend service for an Agent-Based RAG Chatbot, integrating Qdrant retrieval and Gemini Agents SDK with FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.9.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.23.2",
    "google-generativeai>=0.3.1",
    "langchain-google-genai>=0.0.9",
    "langchain-community>=0.0.2"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]